+++ src/paywall-previews/Preview2/Preview2.tsx
@@ -20,39 +20,14 @@ import {
 } from './styled';
 
 type Props = {
-  /**
-   * Branding type
-   */
   branding?: Branding;
-  /**
-   * Minimum width
-   */
   minWidth?: string;
-  /**
-   * Minimum height
-   */
   minHeight?: string;
-  /**
-   * Preview height
-   */
   height?: string;
-  /**
-   * Preview width
-   */
   width?: string;
-  /**
-   * Login footer label
-   */
   loginFooterLabel?: string;
-  /**
-   * Whether the asset is restricted or not
-   */
   isRestrictedAsset?: boolean;
-  /**
-   * Function which handles the modal state changes
-   */
   handleOpenModal?: (e: any) => any;
-  // premium content label
   premiumContentLabel?: string;
 };