+++ src/paywall-previews/Preview5/Preview5.tsx
@@ -25,35 +25,13 @@ import {
 } from './styled';
 
 type Props = {
-  /**
-   * Branding type
-   */
   branding?: Branding;
-  /**
-   * Minimum width
-   */
   minWidth?: string;
-  /**
-   * Minimum height
-   */
   minHeight?: string;
-  /**
-   * Preview height
-   */
   height?: string;
-  /**
-   * Preview width
-   */
   width?: string;
-  /**
-   * Whether the asset is restricted or not
-   */
   isRestrictedAsset?: boolean;
-  /**
-   * Function which handles the modal state changes
-   */
   handleOpenModal?: (e: any) => any;
-  // premium content label
   premiumContentLabel?: string;
 };